Google_translate/requirements.txt
PyQt6>=6.0.0
google-cloud-translate>=3.11.0
